+++ src/index.html
@@ -9,6 +9,6 @@
   <title>Virtual-keyboard</title>
 </head>
 <body>
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
